[Fix] 전체적으로 화면 전환시 생기는 탭바 활성화/비활성화 구현
- extension을 이용해 코드 가독성 높이기 구현
- myScoreViewController의 코드 길이가 너무 길어 myScoreViewController에서 사용하는 Tooltip Class를 따로 파일 만들어서 빼뒀습니다.
