+++ PARD/Sources/HamburgerBarViewController.swift
@@ -40,6 +40,7 @@ class HamburgerBarViewController: UIViewController {
         }
         
     }
+    
     @objc private func didTapedDimmedView() {
         self.dismiss(animated: true) {
             self.didDismiss?()
